refactor: move `DROP` and `KEEP` classes into mapper file
