+++ chidian/__init__.py
@@ -2,8 +2,7 @@ from .core import get, put
 from .data_mapping import DataMapping
 from .dict_group import DictGroup
 from .lib.dsl_parser import parse_path_peg as parse_path
-from .lib.types import DROP, KEEP
-from .mapper import Mapper
+from .mapper import DROP, KEEP, Mapper
 from .partials import ChainableFunction, FunctionChain
 
 __all__ = [
